+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## tip
 
-* BUGFIX: fix issue when VictoriaLogs returns different fields without stream labels. See [this issue](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/23). 
+* BUGFIX: fix bug with displaying responses with a custom set of fields. See [this issue](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/23).
 
 ## v0.2.1